requirement.txt
robotframework==3.1.2
robotframework-appiumlibrary==1.5.0.4
robotframework-jsonlibrary==0.3.1
robotframework-pabot==1.2.1
robotframework-selenium2library==3.0.0
